SimpLL: fix bug in DifferentialFunctionComparator.
Convert the values returned by getNumOperands to signed integer before
subtracting them to avoid integer overflow.

+++ diffkemp/simpll/DifferentialFunctionComparator.cpp
@@ -225,8 +225,9 @@ int DifferentialFunctionComparator::cmpOperations(
                     }
 
                     if (Result && config.ControlFlowOnly
-                        && abs(CL->getNumOperands() - CR->getNumOperands())
-                                   == 1) {
+                        && abs(((int)CL->getNumOperands()
+                                - (int)CR->getNumOperands())
+                               == 1)) {
                         needToCmpOperands = false;
                         return cmpCallsWithExtraArg(CL, CR);
                     }
